refactored patreonThread to make it more readable
