--- src/main/java/PDA/selenium/WebScraper.java
@@ -1,4 +1,0 @@
-package PDA.selenium;
-
-public class WebScraper {
-}
